+++ magma/lib/magma/query/predicate.rb
@@ -378,7 +378,8 @@ require_relative 'predicate/terminal'
 require_relative 'predicate/metrics'
 require_relative 'predicate/table'
 require_relative 'predicate/matrix'
-require_relative 'predicate/subquery'
+require_relative 'predicate/subquery_filter'
+require_relative 'predicate/subquery_predicate'
 require_relative 'predicate/subquery_utils'
 require_relative 'subquery_inner'
 require_relative 'subquery_outer'
